+++ src/Controller/ProductController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 
+use ApiPlatform\Core\Api\IriConverterInterface;
 use App\Entity\Category;
 use App\Entity\Product;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -22,9 +23,14 @@ class ProductController extends AbstractController
     /**
      * @Route("/category/{id}", name="app_category")
      */
-    public function showCategory(Category $category): Response
-    {
-        return $this->render('product/index.html.twig');
+    public function showCategory(
+        Category $category, // by id we symfony can find the entity and inject the Category (entity - not integer ID)
+        IriConverterInterface $iriConverter
+    ): Response {
+        return $this->render('product/index.html.twig', [
+            // pass category IRI (not integer ID)
+            'currentCategoryId' => $iriConverter->getIriFromItem($category),
+        ]);
     }
 
     /**